add figure shortcode (#51)
The figure element is the currently recommended way to include
self-contained flow content such as images, charts, videos and other
visual elements. It allows for the inclusion of captions. It is used by
several css frameworks in handling images and other media, where the
classes are mainly applied to figure elements instead. Including it in
the core shortcodes plugin would be very convenient.

Signed-off-by: Gerard Salvatella <gerard.salvatella@protonmail.com>